Remove old boilerplate code
