+++ dimod/__init__.py
@@ -43,7 +43,7 @@ import dimod.generators
 from dimod.exceptions import *
 import dimod.exceptions
 
-from dimod.higherorder import make_quadratic, poly_energy, poly_energies
+from dimod.higherorder import make_quadratic, poly_energy, poly_energies, BinaryPolynomial
 import dimod.higherorder
 
 from dimod.package_info import __version__, __author__, __authoremail__, __description__
